+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
 -   Simplifying event system.
 -   Applying values in `animate.transitionEnd` if not initial animation.
+-   Made drag constraints only apply if a value isn't animating.
 
 ## [1.5.0] 2019-08-02